fix: error save position_iteration_ts

+++ src/main/resources/db/migration/V2021_12_17_0945.sql
@@ -1,5 +1,5 @@
 ALTER TABLE objects
-    ADD position_iteration_ts int null;
+    ADD position_iteration_ts decimal(21) null;
 
 ALTER TABLE objects
     ADD position_iteration int null;
